requirements.txt
streamlit>=1.38
pandas>=2.0
python-docx>=0.8.11
xlsxwriter>=3.2
requests>=2.28
